requirements.txt
python-docx==1.2.0
PyMuPDF==1.26.7
pillow==12.1.0